requirements.txt
pymongo==4.7.2
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
